requirements.txt
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
openai==1.3.5
flask-sqlalchemy==3.1.1
sqlalchemy==2.0.23
